+++ src/config.rs
@@ -8,6 +8,10 @@ use serde::{Deserialize, Serialize};
 use crate::error::FatalError;
 
 pub trait ConfigSource {
+    fn exclude_paths(&self) -> Option<&[String]> {
+        None
+    }
+
     fn sign_commit(&self) -> Option<bool> {
         None
     }
@@ -90,6 +94,7 @@ pub trait ConfigSource {
 #[serde(deny_unknown_fields, default)]
 #[serde(rename_all = "kebab-case")]
 pub struct Config {
+    pub exclude_paths: Option<Vec<String>>,
     pub sign_commit: Option<bool>,
     pub push_remote: Option<String>,
     pub disable_publish: Option<bool>,
@@ -114,6 +119,9 @@ pub struct Config {
 
 impl Config {
     pub fn update(&mut self, source: &dyn ConfigSource) {
+        if let Some(exclude_paths) = source.exclude_paths() {
+            self.exclude_paths = Some(exclude_paths.to_vec());
+        }
         if let Some(sign_commit) = source.sign_commit() {
             self.sign_commit = Some(sign_commit);
         }
@@ -175,6 +183,10 @@ impl Config {
         if let Some(dependent_version) = source.dependent_version() {
             self.dependent_version = Some(dependent_version);
         }
+    }
+
+    pub fn exclude_paths(&self) -> Option<&[String]> {
+        self.exclude_paths.as_ref().map(|v| v.as_ref())
     }
 
     pub fn sign_commit(&self) -> bool {
@@ -279,6 +291,10 @@ impl Config {
 }
 
 impl ConfigSource for Config {
+    fn exclude_paths(&self) -> Option<&[String]> {
+        self.exclude_paths.as_ref().map(|v| v.as_ref())
+    }
+
     fn sign_commit(&self) -> Option<bool> {
         self.sign_commit
     }
@@ -412,10 +428,40 @@ struct CargoManifest {
     package: Option<CargoPackage>,
 }
 
+impl CargoManifest {
+    fn into_config(self) -> Option<Config> {
+        self.package.and_then(|p| p.into_config())
+    }
+}
+
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
 #[serde(default)]
 struct CargoPackage {
+    include: Option<Vec<String>>,
+    exclude: Option<Vec<String>>,
     metadata: Option<CargoMetadata>,
+}
+
+impl CargoPackage {
+    fn into_config(self) -> Option<Config> {
+        if self.include.is_none() && self.exclude.is_none() && self.metadata.is_none() {
+            return None;
+        }
+        let CargoPackage {
+            include,
+            exclude,
+            metadata,
+        } = self;
+        let mut config = metadata.and_then(|m| m.release).unwrap_or_default();
+        if config.exclude_paths.is_none() {
+            if let Some(_include) = include {
+                log::trace!("Ignoring `include` from Cargo.toml");
+            } else if let Some(exclude) = exclude {
+                config.exclude_paths = Some(exclude);
+            }
+        }
+        Some(config)
+    }
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
@@ -435,7 +481,7 @@ fn get_config_from_manifest(manifest_path: &Path) -> Result<Option<Config>, Fata
     if manifest_path.exists() {
         let m = load_from_file(manifest_path).map_err(FatalError::from)?;
         let c: CargoManifest = toml::from_str(&m).map_err(FatalError::from)?;
-        Ok(c.package.and_then(|p| p.metadata).and_then(|m| m.release))
+        Ok(c.into_config())
     } else {
         Ok(None)
     }